Add logging and class for Parser
